refactoring local-auth
