+++ centreon/tests/e2e/cypress/e2e/Authentication/Login/02-local-authentication.feature
